Split DerivedStore off from LocalStore
