+++ crates/dashchat-node/src/testing/test_node.rs
@@ -14,8 +14,12 @@ use mailbox_client::MailboxClient;
 
 use crate::{
     AgentId, DeviceGroupPayload, NodeConfig, Notification, Payload, Profile,
-    filesystem::Filesystem, mailbox::MailboxOperation, node::Node, stores::LocalStore,
-    testing::TestMailbox, testing::behavior::Behavior, topic::TopicId,
+    filesystem::Filesystem,
+    mailbox::MailboxOperation,
+    node::Node,
+    stores::{DerivedStore, LocalStore},
+    testing::{TestMailbox, behavior::Behavior},
+    topic::TopicId,
 };
 
 #[derive(Clone, derive_more::Deref, derive_more::Debug)]
@@ -37,9 +41,11 @@ impl TestNode {
         let (notification_tx, notification_rx) = tokio::sync::mpsc::channel(100);
 
         let filesystem = Filesystem::new(dir.path().to_path_buf());
-        let local_store = LocalStore::new(filesystem.local_store_path())
+        let pool = crate::stores::create_sqlite_pool(filesystem.local_store_path())
             .await
             .unwrap();
+        let local_store = LocalStore::new(pool.clone()).await.unwrap();
+        let _reducer = DerivedStore::new(pool.clone()).await.unwrap();
         if config.use_alias {
             local_store.device_id().await.unwrap().alias_named(name);
             local_store.agent_id().await.unwrap().alias_named(name);
@@ -84,9 +90,11 @@ impl TestNode {
         let (notification_tx, notification_rx) = tokio::sync::mpsc::channel(100);
 
         let filesystem = Filesystem::new(store_dir.path().to_path_buf());
-        let local_store = LocalStore::new(filesystem.local_store_path())
+        let pool = crate::stores::create_sqlite_pool(filesystem.local_store_path())
             .await
             .unwrap();
+        let local_store = LocalStore::new(pool.clone()).await.unwrap();
+        let _reducer = DerivedStore::new(pool).await.unwrap();
         local_store.device_id().await.unwrap().alias_named(name);
         local_store.agent_id().await.unwrap().alias_named(name);
         drop(local_store);
